requirement.txt
torchvision==0.21.0
albumentations==1.4.11
numpy==1.24.4
scipy==1.10.1
opencv-python==4.10.0.84
patchify==0.2.3
pillow==10.2.0
opensimplex==0.4.5.1
noise==1.2.2
imgaug==0.4.0
einops==0.8.0